Removed import statements that weren't used. Moved code into src folder. Properly indented InsightTester.java.
